moving _traitsui_ui into _ui_tester_register
