+++ traitsui/testing/tester/_ui_tester_register/qt4/_traitsui/instance_editor.py
@@ -13,7 +13,7 @@ from traitsui.testing.tester import command
 from traitsui.testing.tester._ui_tester_register.qt4._interaction_helpers import (  # noqa
     mouse_click_qwidget
 )
-from traitsui.testing.tester._traitsui_ui import (
+from traitsui.testing.tester._ui_tester_register._traitsui_ui import (
     register_traitsui_ui_solvers,
 )
 from traitsui.qt4.instance_editor import (
