+++ traitsui/testing/tester/_ui_tester_register/qt4/_traitsui/list_editor.py
@@ -12,7 +12,7 @@ from traitsui.testing.tester import command, locator
 from traitsui.testing.tester._ui_tester_register._common_ui_targets import (
     _BaseSourceWithLocation
 )
-from traitsui.testing.tester._traitsui_ui import (
+from traitsui.testing.tester._ui_tester_register._traitsui_ui import (
     register_traitsui_ui_solvers,
 )
 from traitsui.testing.tester._ui_tester_register.qt4 import (
